refactor: update package name and groupId
