+++ app/pom.xml
@@ -2,7 +2,7 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.github.halimath.fatesessionboard</groupId>
+        <groupId>com.github.halimath.fatecoreremotetable</groupId>
         <artifactId>parent</artifactId>
         <version>0.1.0</version>
         <relativePath>..</relativePath>
