+++ service/src/main/java/com/github/halimath/fatecoreremotetable/entity/Table.java
@@ -1,4 +1,4 @@
-package com.github.halimath.fatetable.entity;
+package com.github.halimath.fatecoreremotetable.entity;
 
 import java.util.HashSet;
 import java.util.Optional;
