--- service/src/main/java/com/github/halimath/fatetable/entity/CommunicationParticipant.java
@@ -1,5 +1,0 @@
-package com.github.halimath.fatetable.entity;
-
-public interface CommunicationParticipant {
-    
-}
